+++ src/io/mod.rs
@@ -27,10 +27,6 @@ define_whichever! {
     impl Read for ReadFileZip {}
 }
 
-/// A type alias for the `FastQReader` used by IRMA-core's trimmer and
-/// preprocess
-type FastQReaderIc = FastQReader<ReadFileZip>;
-
 define_whichever! {
     #[doc="An enum for the different acceptable output types"]
     #[derive(Debug)]
@@ -57,6 +53,63 @@ define_whichever! {
     impl Read for ReadFileStdin {}
 }
 
+/// Uses [`readers_from_files`] and [`RecordWriters::from_filename`] to generate
+/// readers from the input paths and writers for the output paths.
+///
+/// ## Errors
+///
+/// In addition to any IO errors returned by the above functions, if any of the
+/// input paths equal any of the output paths, then an error is returned.
+#[inline]
+pub(crate) fn get_paired_readers_and_writers<R: RecordReader<ReadFileZip>>(
+    input1: impl AsRef<Path>, input2: Option<impl AsRef<Path>>, output1: Option<impl AsRef<Path>>,
+    output2: Option<impl AsRef<Path>>,
+) -> std::io::Result<(R, Option<R>, RecordWriters<WriteFileZipStdout>, IoThreads)> {
+    fn inner<R: RecordReader<ReadFileZip>>(
+        input1: &Path, input2: Option<&Path>, output1: Option<&Path>, output2: Option<&Path>,
+    ) -> std::io::Result<(R, Option<R>, RecordWriters<WriteFileZipStdout>, IoThreads)> {
+        if let Some(output1) = output1 {
+            if input1 == output1 {
+                return Err(std::io::Error::other(
+                    "The first input file is the same as the first output file",
+                ));
+            } else if let Some(input2) = input2
+                && input2 == output1
+            {
+                return Err(std::io::Error::other(
+                    "The second input file is the same as the first output file",
+                ));
+            }
+        }
+
+        if let Some(output2) = output2 {
+            if input1 == output2 {
+                return Err(std::io::Error::other(
+                    "The first input file is the same as the second output file",
+                ));
+            } else if let Some(input2) = input2
+                && input2 == output2
+            {
+                return Err(std::io::Error::other(
+                    "The second input file is the same as the second output file",
+                ));
+            }
+        }
+
+        let (reader1, reader2, threads) =
+            readers_from_files::<R, _>(input1, input2).map_err(Into::<std::io::Error>::into)?;
+        let writer = RecordWriters::from_filename(output1, output2)?;
+        Ok((reader1, reader2, writer, threads))
+    }
+
+    inner(
+        input1.as_ref(),
+        input2.as_ref().map(AsRef::as_ref),
+        output1.as_ref().map(AsRef::as_ref),
+        output2.as_ref().map(AsRef::as_ref),
+    )
+}
+
 /// Opens a single FASTQ file.
 ///
 /// If the filename ends in `gz`, a thread is spawned with [`spawn_decoder`] to
@@ -78,63 +131,41 @@ pub(crate) fn is_gz<P: AsRef<Path>>(path: P) -> bool {
     path.as_ref().extension().is_some_and(|ext| ext == "gz")
 }
 
-/// Open a single FASTQ file.
+/// Creates a reader from a single file.
 ///
-/// If it ends in `gz`, return a [`FastQReader`] backed by
-/// [`ReadFileZip::Zipped`], and return the thread handle for error propagation.
-/// Otherwise, return a [`FastQReader`] backed by [`ReadFileZip::File`].
+/// If it ends in `gz`, return a reader backed by [`ReadFileZip::Zipped`], and
+/// return the thread handle for error propagation. Otherwise, returns a reader
+/// backed by [`ReadFileZip::File`].
 #[inline]
-pub(crate) fn open_fastq_file<P: AsRef<Path>>(path: P) -> std::io::Result<(FastQReaderIc, Option<IoThread>)> {
+pub(crate) fn reader_from_file<R: RecordReader<ReadFileZip>, P: AsRef<Path>>(
+    path: P,
+) -> std::io::Result<(R, Option<IoThread>)> {
     let file = File::open(&path)?;
 
     if is_gz(&path) {
         let (pipe, thread) = spawn_decoder(path)?;
-        Ok((FastQReader::from_readable(ReadFileZip::Zipped(pipe))?, Some(thread)))
+        Ok((R::from_readable(ReadFileZip::Zipped(pipe))?, Some(thread)))
     } else {
-        Ok((FastQReader::from_readable(ReadFileZip::File(file))?, None))
+        Ok((R::from_readable(ReadFileZip::File(file))?, None))
     }
 }
 
-/// Open one or two FASTQ files. The type of data backing the [`FastQReader`]
-/// depends on the output of [`open_fastq_file`]. The thread handles are grouped
-/// together in [`IoThreads`], if either file was zipped.
-///
-/// ## Errors
-///
-/// `path1` and `path2` must both exist and contain FASTQ data (if not `None`).
-/// Also, if either is zipped, the creation of the pipe must succeed.
-/// Open a single FASTQ/FASTA file.
-///
-/// If it ends in `gz`, return a [`FastQReader`] backed by
-/// [`ReadFileZip::Zipped`], and return the thread handle for error propagation.
-/// Otherwise, return a [`FastQReader`] backed by [`ReadFileZip::File`].
+/// Creates one or two readers from one or two files using the strategy given by
+/// [`reader_from_file`]. The thread handles are grouped together in
+/// [`IoThreads`].
 #[inline]
-pub(crate) fn open_fastx_file<P: AsRef<Path>>(path: P) -> std::io::Result<(FastXReader<ReadFileZip>, Option<IoThread>)> {
-    let file = File::open(&path)?;
-
-    if is_gz(&path) {
-        let (pipe, thread) = spawn_decoder(path)?;
-        Ok((FastXReader::from_readable(ReadFileZip::Zipped(pipe))?, Some(thread)))
-    } else {
-        Ok((FastXReader::from_readable(ReadFileZip::File(file))?, None))
-    }
-}
-
-/// Open one or two FASTQ files using the strategy given by [`open_fastq_file`].
-/// The thread handles are grouped together in [`IoThreads`].
-#[inline]
-pub(crate) fn open_fastq_files<P: AsRef<Path>>(
+pub(crate) fn readers_from_files<R: RecordReader<ReadFileZip>, P: AsRef<Path>>(
     path1: P, path2: Option<P>,
-) -> Result<(FastQReaderIc, Option<FastQReaderIc>, IoThreads), OpenFastqError> {
+) -> Result<(R, Option<R>, IoThreads), ReaderFromFileError> {
     let Some(path2) = path2 else {
-        let (reader, thread) = open_fastq_file(&path1).map_err(|e| OpenFastqError::file1(&path1, e))?;
+        let (reader, thread) = reader_from_file(&path1).map_err(|e| ReaderFromFileError::file1(&path1, e))?;
         let threads = IoThreads(thread, None);
 
         return Ok((reader, None, threads));
     };
 
-    let (reader1, thread1) = open_fastq_file(&path1).map_err(|e| OpenFastqError::file1(&path1, e))?;
-    let (reader2, thread2) = open_fastq_file(&path2).map_err(|e| OpenFastqError::file2(&path2, e))?;
+    let (reader1, thread1) = reader_from_file(&path1).map_err(|e| ReaderFromFileError::file1(&path1, e))?;
+    let (reader2, thread2) = reader_from_file(&path2).map_err(|e| ReaderFromFileError::file2(&path2, e))?;
     let threads = IoThreads(thread1, thread2);
     Ok((reader1, Some(reader2), threads))
 }
@@ -142,37 +173,13 @@ pub(crate) fn open_fastq_files<P: AsRef<Path>>(
 /// Creates a [`WriteFileZipStdout`], using `path` to determine whether a
 /// regular file, zipped file, or stdout should be used.
 ///
-/// ## Errors
-///
-/// Creation of `path` must be successful, if a path is specified.
-/// Open one or two FASTQ/FASTA files using the strategy given by
-/// [`open_fastx_file`]. The thread handles are grouped together in
-/// [`IoThreads`].
-#[inline]
-pub(crate) fn open_fastx_files<P: AsRef<Path>>(
-    path1: P, path2: Option<P>,
-) -> std::io::Result<(FastXReader<ReadFileZip>, Option<FastXReader<ReadFileZip>>, IoThreads)> {
-    let Some(path2) = path2 else {
-        let (reader, thread) = open_fastx_file(path1)?;
-        let threads = IoThreads(thread, None);
-
-        return Ok((reader, None, threads));
-    };
-
-    let (reader1, thread1) = open_fastx_file(path1)?;
-    let (reader2, thread2) = open_fastx_file(path2)?;
-    let threads = IoThreads(thread1, thread2);
-    Ok((reader1, Some(reader2), threads))
-}
-
-/// Creates a [`WriteFileZipStdout`], using `path` to determine whether a regular
-/// file, zipped file, or stdout should be used.
+/// The path is included in the error if the file cannot be created.
 #[inline]
 pub(crate) fn create_writer<P: AsRef<Path>>(path: Option<P>) -> std::io::Result<WriteFileZipStdout> {
     let writer = match path {
         Some(ref p) => {
             let is_gz = p.as_ref().extension().is_some_and(|ext| ext == "gz");
-            let file = File::create(p)?;
+            let file = File::create(p).map_failed_write(p.as_ref())?;
             let buf_writer = BufWriter::new(file);
 
             if is_gz {
@@ -238,22 +245,22 @@ impl IoThreads {
 /// occurred with the first file or second file specified.
 #[non_exhaustive]
 #[derive(Debug)]
-pub enum OpenFastqError {
+pub enum ReaderFromFileError {
     File1 { path: PathBuf, source: std::io::Error },
     File2 { path: PathBuf, source: std::io::Error },
 }
 
-impl std::fmt::Display for OpenFastqError {
+impl std::fmt::Display for ReaderFromFileError {
     #[inline]
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match self {
-            OpenFastqError::File1 { path, source } => {
+            ReaderFromFileError::File1 { path, source } => {
                 write!(
                     f,
                     "Failed to read the data in first file {path:#?} due to the error:\n{source}"
                 )
             }
-            OpenFastqError::File2 { path, source } => {
+            ReaderFromFileError::File2 { path, source } => {
                 write!(
                     f,
                     "Failed to read the data in second file {path:#?} due to the error:\n{source}"
@@ -265,8 +272,8 @@ impl std::fmt::Display for OpenFastqError {
 /// A struct containing two writers for paired reads: one for left reads and one
 /// for the right.
 pub struct PairedWriters<W: Write> {
-    writer1: W,
-    writer2: W,
+    pub(crate) writer1: W,
+    pub(crate) writer2: W,
 }
 
 impl<W: Write> PairedWriters<W> {
@@ -322,32 +329,32 @@ impl RecordWriters<WriteFileZipStdout> {
     /// Creates a [`RecordWriters`] from two files, interpretted as
     /// [`WriteFileZipStdout`] writers.
     pub fn from_filename<P: AsRef<Path>>(writer1: Option<P>, writer2: Option<P>) -> std::io::Result<Self> {
-        let writer1 = create_writer(writer1)?;
+        let writer1 = create_writer(writer1.as_ref())?;
         let writer2 = match writer2 {
-            Some(path) => Some(create_writer(Some(path))?),
+            Some(path) => Some(create_writer(Some(path.as_ref()))?),
             None => None,
         };
         Ok(Self::new(writer1, writer2))
     }
 }
 
-impl Error for OpenFastqError {
+impl Error for ReaderFromFileError {
     #[inline]
     fn source(&self) -> Option<&(dyn Error + 'static)> {
         match self {
-            OpenFastqError::File1 { source, .. } => Some(source),
-            OpenFastqError::File2 { source, .. } => Some(source),
+            ReaderFromFileError::File1 { source, .. } => Some(source),
+            ReaderFromFileError::File2 { source, .. } => Some(source),
         }
     }
 }
 
-impl From<OpenFastqError> for std::io::Error {
-    fn from(error: OpenFastqError) -> Self {
+impl From<ReaderFromFileError> for std::io::Error {
+    fn from(error: ReaderFromFileError) -> Self {
         std::io::Error::other(error)
     }
 }
 
-impl OpenFastqError {
+impl ReaderFromFileError {
     #[inline]
     fn file1(path: impl AsRef<Path>, source: std::io::Error) -> Self {
         Self::File1 {
@@ -366,16 +373,36 @@ impl OpenFastqError {
 }
 
 pub trait MapFailedWriteExt<T> {
-    fn map_failed_write(self, path: Option<&PathBuf>) -> std::io::Result<T>;
+    fn map_failed_write<P: AsRef<Path>>(self, path: P) -> std::io::Result<T>;
 }
 
 impl<T> MapFailedWriteExt<T> for std::io::Result<T> {
-    fn map_failed_write(self, path: Option<&PathBuf>) -> std::io::Result<T> {
+    fn map_failed_write<P: AsRef<Path>>(self, path: P) -> std::io::Result<T> {
         self.map_err(|e| {
             std::io::Error::other(format!(
-                "Failed to open {path:#?} for writing due to the error:\n{e}",
-                path = path.unwrap()
+                "Failed to open {path} for writing due to the error:\n{e}",
+                path = path.as_ref().display()
             ))
         })
+    }
+}
+
+/// A trait unifying [`FastQReader`] and [`FastXReader`].
+pub trait RecordReader<R: Read>: Sized {
+    /// Creates a reader from `read`.
+    fn from_readable(read: R) -> std::io::Result<Self>;
+}
+
+impl<R: Read> RecordReader<R> for FastQReader<R> {
+    #[inline]
+    fn from_readable(read: R) -> std::io::Result<Self> {
+        FastQReader::from_readable(read)
+    }
+}
+
+impl<R: Read> RecordReader<R> for FastXReader<R> {
+    #[inline]
+    fn from_readable(read: R) -> std::io::Result<Self> {
+        FastXReader::from_readable(read)
     }
 }